move lib-specific test to the right place
